Add mutation rate summary to mutation mapper
Determine somatic mutation rate. That is divide number of cases with a somatic
mutation by total number of cases queried. Calculate germline mutation rate if
there are germline mutations. For the impact study there is a specific clinical
attribute that indicates whether the patient was screened for germline
mutations `12_245_PARTC_CONSENTED`. Use number of cases that consented as
denominator if that is available (subset of total).